+++ js/game.js
@@ -3,63 +3,24 @@ var needCanvasUpdate = true;
 var NaNalert = false;
 var gameEnded = false;
 
-
-let modInfo = {
-	name: "The Basic Tree",
-	id: "gapples2",
-	pointsName: "dust",
-	discordName: "",
-	discordLink: "",
-	changelogLink: "https://github.com/Acamaeda/The-Modding-Tree/blob/master/changelog.md",
-	offlineLimit: 1  // In hours
+// Don't change this
+const TMT_VERSION = {
+	tmtNum: "2.1",
+	tmtName: "We should have thought of this sooner!"
 }
-
-// Set your version in num and name, but leave the tmt values so people know what version it is
-let VERSION = {
-	pre: true,
-	num: "1.1.1 (cheapeners are changed)",
-	name: "Cheapening Stuff",
-	tmtNum: "2.0.2",
-	tmtName: "Pinnacle of Achievement Mountain"
-}
-
-// Determines if it should show points/sec
-function canGenPoints(){
-	if (hasUpgrade("b",11)){
-	return true
-	} else {return false}
-}
-
-// Calculate points/sec!
-function getPointGen() {
-	if(!canGenPoints())
-		return new Decimal(0)
-
-	let gain = new Decimal(0)
-	if (hasUpgrade("b", 11)) gain = gain.add(1)
-	if (hasUpgrade("b", 21)) gain = gain.add(1)
-	if (hasUpgrade("b", 31)) gain = gain.add(2)
-	if (hasUpgrade("b", 12)) gain = gain.mul(1.5)
-	if (hasUpgrade("b", 22)) gain = gain.mul(1.5)
-	if (hasUpgrade("b", 32)) gain = gain.mul(2.0)
-	return gain
-}
-
-
-
 function getResetGain(layer, useType = null) {
 	let type = useType
-	if (!useType) type = layers[layer].type
+	if (!useType) type = tmp[layer].type
 	if(tmp[layer].type == "none")
 		return new Decimal (0)
 	if (tmp[layer].gainExp.eq(0)) return new Decimal(0)
 	if (type=="static") {
-		if ((!layers[layer].canBuyMax()) || tmp[layer].baseAmount.lt(tmp[layer].requires)) return new Decimal(1)
-		let gain = tmp[layer].baseAmount.div(tmp[layer].requires).div(tmp[layer].gainMult).max(1).log(layers[layer].base).times(tmp[layer].gainExp).pow(Decimal.pow(layers[layer].exponent, -1))
+		if ((!tmp[layer].canBuyMax) || tmp[layer].baseAmount.lt(tmp[layer].requires)) return new Decimal(1)
+		let gain = tmp[layer].baseAmount.div(tmp[layer].requires).div(tmp[layer].gainMult).max(1).log(tmp[layer].base).times(tmp[layer].gainExp).pow(Decimal.pow(tmp[layer].exponent, -1))
 		return gain.floor().sub(player[layer].points).add(1).max(1);
 	} else if (type=="normal"){
 		if (tmp[layer].baseAmount.lt(tmp[layer].requires)) return new Decimal(0)
-		let gain = tmp[layer].baseAmount.div(tmp[layer].requires).pow(layers[layer].exponent).times(tmp[layer].gainMult).pow(tmp[layer].gainExp)
+		let gain = tmp[layer].baseAmount.div(tmp[layer].requires).pow(tmp[layer].exponent).times(tmp[layer].gainMult).pow(tmp[layer].gainExp)
 		if (gain.gte("e1e7")) gain = gain.sqrt().times("e5e6")
 		return gain.floor().max(0);
 	} else if (type=="custom"){
@@ -71,7 +32,7 @@ function getResetGain(layer, useType = null) {
 
 function getNextAt(layer, canMax=false, useType = null) {
 	let type = useType
-	if (!useType) type = layers[layer].type
+	if (!useType) type = tmp[layer].type
 	if(tmp[layer].type == "none")
 		return new Decimal (Infinity)
 
@@ -80,17 +41,17 @@ function getNextAt(layer, canMax=false, useType = null) {
 
 	if (type=="static") 
 	{
-		if (!layers[layer].canBuyMax()) canMax = false
+		if (!tmp[layer].canBuyMax) canMax = false
 		let amt = player[layer].points.plus((canMax&&tmp[layer].baseAmount.gte(tmp[layer].nextAt))?tmp[layer].resetGain:0)
-		let extraCost = Decimal.pow(layers[layer].base, amt.pow(tmp[layer].exponent).div(tmp[layer].gainExp)).times(tmp[layer].gainMult)
+		let extraCost = Decimal.pow(tmp[layer].base, amt.pow(tmp[layer].exponent).div(tmp[layer].gainExp)).times(tmp[layer].gainMult)
 		let cost = extraCost.times(tmp[layer].requires).max(tmp[layer].requires)
-		if (layers[layer].roundUpCost) cost = cost.ceil()
+		if (tmp[layer].roundUpCost) cost = cost.ceil()
 		return cost;
 	} else if (type=="normal"){
 		let next = tmp[layer].resetGain.add(1)
 		if (next.gte("e1e7")) next = next.div("e5e6").pow(2)
 		next = next.root(tmp[layer].gainExp).div(tmp[layer].gainMult).root(tmp[layer].exponent).times(tmp[layer].requires).max(tmp[layer].requires)
-		if (layers[layer].roundUpCost) next = next.ceil()
+		if (tmp[layer].roundUpCost) next = next.ceil()
 		return next;
 	} else if (type=="custom"){
 		return layers[layer].getNextAt(canMax)
@@ -100,7 +61,7 @@ function getNextAt(layer, canMax=false, useType = null) {
 
 // Return true if the layer should be highlighted. By default checks for upgrades only.
 function shouldNotify(layer){
-	for (id in layers[layer].upgrades){
+	for (id in tmp[layer].upgrades){
 		if (!isNaN(id)){
 			if (canAffordUpgrade(layer, id) && !hasUpgrade(layer, id) && tmp[layer].upgrades[id].unlocked){
 				return true
@@ -134,7 +95,7 @@ function rowReset(row, layer) {
 			layers[lr].doReset(layer)
 		}
 		else
-			if(layers[layer].row > layers[lr].row && row !== "side") layerDataReset(lr)
+			if(tmp[layer].row > tmp[lr].row && row !== "side") layerDataReset(lr)
 	}
 }
 
@@ -145,7 +106,6 @@ function layerDataReset(layer, keep = []) {
 		if (player[layer][keep[thing]] !== undefined)
 			storedData[keep[thing]] = player[layer][keep[thing]]
 	}
-	console.log(storedData)
 
 	player[layer] = layers[layer].startData();
 	player[layer].upgrades = []
@@ -178,15 +138,15 @@ function generatePoints(layer, diff) {
 var prevOnReset
 
 function doReset(layer, force=false) {
-	let row = layers[layer].row
+	let row = tmp[layer].row
 	if (!force) {
 		if (tmp[layer].baseAmount.lt(tmp[layer].requires)) return;
 		let gain = tmp[layer].resetGain
-		if (layers[layer].type=="static") {
+		if (tmp[layer].type=="static") {
 			if (tmp[layer].baseAmount.lt(tmp[layer].nextAt)) return;
-			gain =(layers[layer].canBuyMax() ? gain : 1)
+			gain =(tmp[layer].canBuyMax ? gain : 1)
 		} 
-		if (layers[layer].type=="custom") {
+		if (tmp[layer].type=="custom") {
 			if (!tmp[layer].canReset) return;
 		} 
 
@@ -201,8 +161,8 @@ function doReset(layer, force=false) {
 			player[layer].unlocked = true;
 			needCanvasUpdate = true;
 
-			if (layers[layer].increaseUnlockOrder){
-				lrs = layers[layer].increaseUnlockOrder
+			if (tmp[layer].increaseUnlockOrder){
+				lrs = tmp[layer].increaseUnlockOrder
 				for (lr in lrs)
 					if (!player[lrs[lr]].unlocked) player[lrs[lr]].unlockOrder++
 			}
@@ -219,7 +179,7 @@ function doReset(layer, force=false) {
 	}
 
 	prevOnReset = {...player} //Deep Copy
-	player.points = (row == 0 ? new Decimal(0) : new Decimal(10))
+	player.points = (row == 0 ? new Decimal(0) : getStartPoints())
 
 	for (let x = row; x >= 0; x--) rowReset(x, layer)
 	rowReset("side", layer)
@@ -237,10 +197,10 @@ function resetRow(row) {
 	rowReset(row+1, post_layers[0])
 	doReset(pre_layers[0], true)
 	for (let layer in layers) {
-		player[layers[layer]].unlocked = false
-		if (player[layers[layer]].unlockOrder) player[layers[layer]].unlockOrder = 0
+		player[layer].unlocked = false
+		if (player[layer].unlockOrder) player[layer].unlockOrder = 0
 	}
-	player.points = new Decimal(10)
+	player.points = getStartPoints()
 	updateTemp();
 	resizeCanvas();
 }
@@ -264,11 +224,14 @@ function canCompleteChallenge(layer, x)
 {
 	if (x != player[layer].activeChallenge) return
 
-	let challenge = layers[layer].challenges[x]
+	let challenge = tmp[layer].challenges[x]
 
 	if (challenge.currencyInternalName){
 		let name = challenge.currencyInternalName
-		if (challenge.currencyLayer){
+		if (challenge.currencyLocation){
+			return !(challenge.currencyLocation[name].lt(challenge.goal)) 
+		}
+		else if (challenge.currencyLayer){
 			let lr = challenge.currencyLayer
 			return !(player[lr][name].lt(readData(challenge.goal))) 
 		}
@@ -300,12 +263,10 @@ function completeChallenge(layer, x) {
 
 VERSION.withoutName = "v" + VERSION.num + (VERSION.pre ? " Pre-Release " + VERSION.pre : VERSION.pre ? " Beta " + VERSION.beta : "")
 VERSION.withName = VERSION.withoutName + (VERSION.name ? ": " + VERSION.name : "")
-
-
 const ENDGAME = new Decimal("e3");
 
 function gameLoop(diff) {
-	if (player.points.gte(ENDGAME) || gameEnded) gameEnded = 1
+	if (isEndgame() || gameEnded) gameEnded = 1
 
 	if (isNaN(diff)) diff = 0
 	if (gameEnded && !player.keepGoing) {
@@ -313,6 +274,10 @@ function gameLoop(diff) {
 		player.tab = "gameEnded"
 	}
 	if (player.devSpeed) diff *= player.devSpeed
+
+	let limit = maxTickLength()
+	if(diff > limit)
+		diff = limit
 
 	addTime(diff)
 	player.points = player.points.add(tmp.pointGen.times(diff)).max(0)