+++ js/utils.js
@@ -3,7 +3,11 @@
 function exponentialFormat(num, precision) {
 	let e = num.log10().floor()
 	let m = num.div(Decimal.pow(10, e))
-	return m.toStringWithDecimalPlaces(3)+"e"+e.toStringWithDecimalPlaces(0)
+	if(m.toStringWithDecimalPlaces(precision) == 10) {
+		m = new Decimal(1)
+		e = e.add(1)
+	}
+	return m.toStringWithDecimalPlaces(precision)+"e"+e.toStringWithDecimalPlaces(0)
 }
 
 function commaFormat(num, precision) {
@@ -34,7 +38,7 @@ function format(decimal, precision=2) {
 		var slog = decimal.slog()
 		if (slog.gte(1e6)) return "F" + format(slog.floor())
 		else return Decimal.pow(10, slog.sub(slog.floor())).toStringWithDecimalPlaces(3) + "F" + commaFormat(slog.floor(), 0)
-	} else if (decimal.gte("1e1000")) return (Math.floor(decimal.mantissa + 0.01) + ("e"+formatWhole(decimal.log10().floor())))
+	} else if (decimal.gte("1e1000")) return exponentialFormat(decimal, 0)
 	else if (decimal.gte(1e9)) return exponentialFormat(decimal, precision)
 	else if (decimal.gte(1e3)) return commaFormat(decimal, 0)
 	else return commaFormat(decimal, precision)
@@ -79,13 +83,19 @@ function startPlayerBase() {
 		keepGoing: false,
 		hasNaN: false,
 		hideChallenges: false,
-		points: new Decimal(10),
+		points: modInfo.initialStartPoints,
 		subtabs: {},
 	}
 }
 
 function getStartPlayer() {
 	playerdata = startPlayerBase()
+	
+	if (addedPlayerData) {
+		extradata = addedPlayerData()
+		for (thing in extradata)
+			playerdata[thing] = extradata[thing]
+	}
 	for (layer in layers){
 		playerdata[layer] = layers[layer].startData()
 		playerdata[layer].buyables = getStartBuyables(layer)
@@ -350,13 +360,13 @@ function milestoneShown(layer, id) {
 function respecBuyables(layer) {
 	if (!layers[layer].buyables) return
 	if (!layers[layer].buyables.respec) return
-	if (!confirm("Are you sure you want to respec? This will force you to do a \"" + (layers[layer].name ? layers[layer].name : layer) + "\" reset as well!")) return
+	if (!confirm("Are you sure you want to respec? This will force you to do a \"" + (tmp[layer].name ? tmp[layer].name : layer) + "\" reset as well!")) return
 	layers[layer].buyables.respec()
 	updateBuyableTemp(layer)
 }
 
 function canAffordUpgrade(layer, id) {
-	let upg = layers[layer].upgrades[id]
+	let upg = tmp[layer].upgrades[id]
 	let cost = tmp[layer].upgrades[id].cost
 	return canAffordPurchase(layer, upg, cost) 
 }
@@ -409,14 +419,22 @@ function buyableEffect(layer, id){
 	return (tmp[layer].buyables[id].effect)
 }
 
+function clickableEffect(layer, id){
+	return (tmp[layer].clickables[id].effect)
+}
+
 function achievementEffect(layer, id){
 	return (tmp[layer].achievements[id].effect)
 }
 
 function canAffordPurchase(layer, thing, cost) {
+
 	if (thing.currencyInternalName){
 		let name = thing.currencyInternalName
-		if (thing.currencyLayer){
+		if (thing.currencyLocation){
+			return !(thing.currencyLocation[name].lt(cost)) 
+		}
+		else if (thing.currencyLayer){
 			let lr = thing.currencyLayer
 			return !(player[lr][name].lt(cost)) 
 		}
@@ -431,14 +449,18 @@ function canAffordPurchase(layer, thing, cost) {
 
 function buyUpg(layer, id) {
 	if (!player[layer].unlocked) return
-	if (!layers[layer].upgrades[id].unlocked) return
+	if (!tmp[layer].upgrades[id].unlocked) return
 	if (player[layer].upgrades.includes(id)) return
-	let upg = layers[layer].upgrades[id]
+	let upg = tmp[layer].upgrades[id]
 	let cost = tmp[layer].upgrades[id].cost
 
 	if (upg.currencyInternalName){
 		let name = upg.currencyInternalName
-		if (upg.currencyLayer){
+		if (upg.currencyLocation){
+			if (upg.currencyLocation[name].lt(cost)) return
+			upg.currencyLocation[name] = upg.currencyLocation[name].sub(cost)
+		}
+		else if (upg.currencyLayer){
 			let lr = upg.currencyLayer
 			if (player[lr][name].lt(cost)) return
 			player[lr][name] = player[lr][name].sub(cost)
@@ -488,10 +510,12 @@ function clickClickable(layer, id) {
 // Function to determine if the player is in a challenge
 function inChallenge(layer, id){
 	let challenge = player[layer].activeChallenge
-	if (challenge==toNumber(id)) return true
+	if (challenge == null) return
+	id = toNumber(id)
+	if (challenge==id) return true
 
 	if (layers[layer].challenges[challenge].countsAs)
-		return layers[layer].challenges[id].countsAs.includes(id)
+		return tmp[layer].challenges[id].countsAs.includes(id)
 }
 
 // ************ Misc ************
@@ -527,7 +551,7 @@ function nodeShown(layer) {
 }
 
 function layerunlocked(layer) {
-	return LAYERS.includes(layer) && (player[layer].unlocked || (tmp[layer].baseAmount.gte(tmp[layer].requires) && layers[layer].layerShown()))
+	return LAYERS.includes(layer) && (player[layer].unlocked || (tmp[layer].baseAmount.gte(tmp[layer].requires) && tmp[layer].layerShown))
 }
 
 function keepGoing() {
@@ -589,7 +613,7 @@ document.onkeydown = function(e) {
 	let key = e.key
 	if (ctrlDown) key = "ctrl+" + key
 	if (onFocused) return
-	if (ctrlDown && key != "-" && key != "_" && key != "+" && key != "=" && key != "r" && key != "R" && key != "F5") e.preventDefault()
+	if (ctrlDown && hotkeys[key]) e.preventDefault()
 	if(hotkeys[key]){
 		if (player[hotkeys[key].layer].unlocked)
 			hotkeys[key].onPress()
@@ -606,7 +630,7 @@ function prestigeButtonText(layer)
 	if(tmp[layer].type == "normal")
 		return `${ player[layer].points.lt(1e3) ? (tmp[layer].resetDescription !== undefined ? tmp[layer].resetDescription : "Reset for ") : ""}+<b>${formatWhole(tmp[layer].resetGain)}</b> ${tmp[layer].resource} ${tmp[layer].resetGain.lt(100) && player[layer].points.lt(1e3) ? `<br><br>Next at ${ (tmp[layer].roundUpCost ? formatWhole(tmp[layer].nextAt) : format(tmp[layer].nextAt))} ${ tmp[layer].baseResource }` : ""}`
 	else if(tmp[layer].type== "static")
-		return `${tmp[layer].resetDescription !== undefined ? tmp[layer].resetDescription : "Reset for "}+<b>${formatWhole(tmp[layer].resetGain)}</b> ${tmp[layer].resource}<br><br>${player[layer].points.lt(20) ? (tmp[layer].baseAmount.gte(tmp[layer].nextAt)&&(tmp[layer].canBuyMax !== undefined) && tmp[layer].canBuyMax?"Next":"Req") : ""}: ${formatWhole(tmp[layer].baseAmount)} / ${(tmp[layer].roundUpCost ? formatWhole(tmp[layer].nextAtDisp) : format(tmp[layer].nextAtDisp))} ${ tmp[layer].baseResource }		
+		return `${tmp[layer].resetDescription !== undefined ? tmp[layer].resetDescription : "Reset for "}+<b>${formatWhole(tmp[layer].resetGain)}</b> ${tmp[layer].resource}<br><br>${player[layer].points.lt(30) ? (tmp[layer].baseAmount.gte(tmp[layer].nextAt)&&(tmp[layer].canBuyMax !== undefined) && tmp[layer].canBuyMax?"Next:":"Req:") : ""} ${formatWhole(tmp[layer].baseAmount)} / ${(tmp[layer].roundUpCost ? formatWhole(tmp[layer].nextAtDisp) : format(tmp[layer].nextAtDisp))} ${ tmp[layer].baseResource }		
 		`
 	else if(tmp[layer].type == "none")
 		return ""